requirements.txt
# django for django workflows
Django>=3.1.7
gunicorn
django-heroku
sqlite-utils>=3.1
django-allauth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
json2html